+++ internal/impl/confluent/serde_json_test.go
@@ -110,7 +110,7 @@ func TestResolveJsonSchema(t *testing.T) {
 			encoder, err := newSchemaRegistryEncoder(urlStr, noopReqSign, nil, subj, true, time.Minute*10, time.Minute, service.MockResources())
 			require.NoError(t, err)
 
-			decoder, err := newSchemaRegistryDecoder(urlStr, noopReqSign, nil, true, service.MockResources())
+			decoder, err := newSchemaRegistryDecoder(urlStr, noopReqSign, nil, decodingConfig{}, service.MockResources())
 			require.NoError(t, err)
 
 			t.Cleanup(func() {